refactor(taskflow): extract feature.boardlist
Move BoardListModel, BoardListActions (LoadBoardListRequested/Succeeded/Failed,
CreateBoard), boardListReducer, DEFAULT_BOARD_COLOR, BoardListScreen, and
BoardSummaryCard into the new feature.boardlist package. Update all consumers
(EffectsMiddleware, AccountStore, App, BoardScreen, ProfileScreen, tests) to
import from the new location.
