+++ examples/taskflow/composeApp/src/commonMain/kotlin/org/reduxkotlin/sample/taskflow/App.kt
@@ -50,6 +50,7 @@ import org.reduxkotlin.sample.taskflow.feature.account.LoginScreen
 import org.reduxkotlin.sample.taskflow.feature.account.ProfileScreen
 import org.reduxkotlin.sample.taskflow.feature.account.StartLogin
 import org.reduxkotlin.sample.taskflow.feature.account.SwitcherScreen
+import org.reduxkotlin.sample.taskflow.feature.boardlist.BoardListScreen
 import org.reduxkotlin.sample.taskflow.feature.settings.SettingsScreen
 import org.reduxkotlin.sample.taskflow.infra.SeedData
 import org.reduxkotlin.sample.taskflow.infra.data.local.LocalStore
@@ -68,7 +69,6 @@ import org.reduxkotlin.sample.taskflow.ui.PredictiveBackHandler
 import org.reduxkotlin.sample.taskflow.ui.adaptive.widthSizeClass
 import org.reduxkotlin.sample.taskflow.ui.components.AdaptiveNav
 import org.reduxkotlin.sample.taskflow.ui.image.initCoil
-import org.reduxkotlin.sample.taskflow.ui.screens.BoardListScreen
 import org.reduxkotlin.sample.taskflow.ui.screens.BoardScreen
 import org.reduxkotlin.sample.taskflow.ui.screens.CardDetailScreen
 import org.reduxkotlin.sample.taskflow.ui.theme.TaskFlowTheme
