+++ examples/taskflow/composeApp/src/commonMain/kotlin/org/reduxkotlin/sample/taskflow/feature/boardlist/BoardListModels.kt
@@ -1,4 +1,4 @@
-package org.reduxkotlin.sample.taskflow.model
+package org.reduxkotlin.sample.taskflow.feature.boardlist
 
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.PersistentMap
@@ -6,11 +6,6 @@ import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.persistentMapOf
 import org.reduxkotlin.sample.taskflow.core.BoardId
 import org.reduxkotlin.sample.taskflow.core.BoardSummary
-
-// --- Per-account models (ALL declared up front; board slices start empty/sentinel) ---
-// Identity (name/email/avatar) is NOT duplicated here — it lives once in CollaboratorsModel
-// (which includes self). SessionModel moved to …feature.account.AccountModels.
-// CollaboratorsModel moved to …feature.collaborators.CollaboratorsModels.
 
 /** Holds the cached board tiles and their display order for one account. */
 public data class BoardListModel(
