+++ examples/taskflow/composeApp/src/commonMain/kotlin/org/reduxkotlin/sample/taskflow/feature/boardlist/BoardSummaryCard.kt
@@ -1,4 +1,4 @@
-package org.reduxkotlin.sample.taskflow.ui.components
+package org.reduxkotlin.sample.taskflow.feature.boardlist
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
