+++ examples/taskflow/composeApp/src/commonMain/kotlin/org/reduxkotlin/sample/taskflow/reducer/AccountReducers.kt
@@ -2,26 +2,18 @@ package org.reduxkotlin.sample.taskflow.reducer
 
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.persistentListOf
-import kotlinx.collections.immutable.persistentMapOf
-import kotlinx.collections.immutable.toPersistentList
 import org.reduxkotlin.sample.taskflow.action.Back
 import org.reduxkotlin.sample.taskflow.action.CancelCreateCard
 import org.reduxkotlin.sample.taskflow.action.CloseCard
-import org.reduxkotlin.sample.taskflow.action.CreateBoard
 import org.reduxkotlin.sample.taskflow.action.EnterEditMode
-import org.reduxkotlin.sample.taskflow.action.LoadBoardListSucceeded
 import org.reduxkotlin.sample.taskflow.action.Navigate
 import org.reduxkotlin.sample.taskflow.action.OpenCard
 import org.reduxkotlin.sample.taskflow.action.StartCreateCard
 import org.reduxkotlin.sample.taskflow.core.Action
-import org.reduxkotlin.sample.taskflow.core.BoardId
-import org.reduxkotlin.sample.taskflow.core.BoardSummary
 import org.reduxkotlin.sample.taskflow.core.NavModel
 import org.reduxkotlin.sample.taskflow.core.Route
-import org.reduxkotlin.sample.taskflow.model.BoardListModel
 
-/** Fallback tile color for a newly created board (Material 3 primary). */
-public const val DEFAULT_BOARD_COLOR: Long = 0xFF6750A4
+// DEFAULT_BOARD_COLOR moved to …feature.boardlist.BoardListReducer
 
 /**
  * Pure per-account reducer for the [NavModel] slice (the navigation stack).
@@ -118,45 +110,5 @@ private fun NavModel.flipCardDetailMode(to: Route.CardDetail.Mode, requireOther:
 }
 
 // sessionReducer moved to …feature.account.AccountReducers
-
-/**
- * Pure per-account reducer for the [BoardListModel] slice (board tile cache + display order).
- *
- * [LoadBoardListSucceeded] replaces the cache, keying tiles by id and taking the DB-aggregate
- * counts as-is. [CreateBoard] appends a fresh tile using the action's pre-minted timestamp (never
- * a reducer-side clock). Returns the same [model] instance unchanged for actions it does not handle.
- *
- * @param model the current board-list slice.
- * @param action the dispatched action.
- * @return the next board-list slice, or [model] unchanged when [action] is not handled.
- */
-public fun boardListReducer(model: BoardListModel, action: Action): BoardListModel = when (action) {
-    is LoadBoardListSucceeded -> {
-        val boards = persistentMapOf<BoardId, BoardSummary>().builder().apply {
-            action.summaries.forEach { put(it.id, it) }
-        }.build()
-        model.copy(
-            boards = boards,
-            order = action.summaries.map { it.id }.toPersistentList(),
-        )
-    }
-
-    is CreateBoard -> model.copy(
-        boards = model.boards.put(
-            action.boardId,
-            BoardSummary(
-                id = action.boardId,
-                name = action.name,
-                color = DEFAULT_BOARD_COLOR,
-                cardCount = 0,
-                doneCount = 0,
-                updatedAt = action.now,
-            ),
-        ),
-        order = model.order.add(action.boardId),
-    )
-
-    else -> model
-}
-
+// boardListReducer moved to …feature.boardlist.BoardListReducer
 // collaboratorsReducer moved to …feature.collaborators.CollaboratorsReducer
