+++ examples/taskflow/composeApp/src/commonMain/kotlin/org/reduxkotlin/sample/taskflow/ui/screens/BoardScreen.kt
@@ -61,11 +61,11 @@ import org.reduxkotlin.sample.taskflow.core.ActivityEntry
 import org.reduxkotlin.sample.taskflow.core.CardId
 import org.reduxkotlin.sample.taskflow.core.ColumnId
 import org.reduxkotlin.sample.taskflow.feature.activity.ActivityModel
+import org.reduxkotlin.sample.taskflow.feature.boardlist.BoardListModel
 import org.reduxkotlin.sample.taskflow.feature.collaborators.CollaboratorsModel
 import org.reduxkotlin.sample.taskflow.feature.undo.Redo
 import org.reduxkotlin.sample.taskflow.feature.undo.Undo
 import org.reduxkotlin.sample.taskflow.feature.undo.UndoModel
-import org.reduxkotlin.sample.taskflow.model.BoardListModel
 import org.reduxkotlin.sample.taskflow.model.BoardModel
 import org.reduxkotlin.sample.taskflow.model.FilterModel
 import org.reduxkotlin.sample.taskflow.model.SyncModel
